Refactor cast_storage
